+++ src/io/hardware/i2c/I2CWrite.java
@@ -65,7 +65,7 @@ public class I2CWrite implements I2COp{
         return true;
     }
     @Override
-    public ArrayList<Integer> doOperation(ExtI2CDevice device) {
+    public ArrayList<Double> doOperation(ExtI2CDevice device) {
         if( device.isDebug())
             Logger.info(device.id()+"(i2c) -> Writing "+Tools.fromBytesToHexString(towrite));
         try{
